Moving core classes (i.e. Controller, Connection, Render, Table/Model) to App/Core folder/namespace
