+++ app/Controllers/QueueController.php
@@ -2,8 +2,8 @@
 
 namespace App\Controllers;
 
+use App\Core\Controller;
 use App\Models\Offer;
-use App\Models\User;
 
 class QueueController extends Controller
 {
