+++ app/Core/Controller.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controllers;
+namespace App\Core;
 
 use App\Models\Category;
 use App\Models\User;
