+++ app/Core/Table.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models;
+namespace App\Core;
 
 class Table extends Connection
 {
